+++ tests/cases/pending/plan.py
@@ -1,4 +1,16 @@
 #!/usr/bin/env python
 from stepup.core.api import step
 
-step("cp -v first.txt second.txt", inp="first.txt", out="second.txt", vol="third.txt")
+info = step("cp -v first.txt second.txt", inp="first.txt", out="second.txt", vol="third.txt")
+
+# Tests for the info object, only useful for testing
+if info.inp != ["first.txt"]:
+    raise AssertionError("Incorrect info.inp")
+if info.env != []:
+    raise AssertionError("Incorrect info.env")
+if info.out != ["second.txt"]:
+    raise AssertionError("Incorrect info.out")
+if info.vol != ["third.txt"]:
+    raise AssertionError("Incorrect info.vol")
+if list(info.filter_vol("*.txt")) != ["third.txt"]:
+    raise AssertionError("Wrong info.filter_vol")